Move vm-specific classes into Debugger
