--- java/src/processing/mode/java/pdex/VMEventListener.java
@@ -1,39 +1,0 @@
-/* -*- mode: java; c-basic-offset: 2; indent-tabs-mode: nil -*- */
-
-/*
-Part of the Processing project - http://processing.org
-Copyright (c) 2012-15 The Processing Foundation
-
-This program is free software; you can redistribute it and/or modify
-it under the terms of the GNU General Public License version 2
-as published by the Free Software Foundation.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program; if not, write to the Free Software Foundation, Inc.
-51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
-*/
-
-package processing.mode.java.pdex;
-
-import com.sun.jdi.event.EventSet;
-
-/**
- * Interface for VM callbacks.
- *
- * @author Martin Leopold <m@martinleopold.com>
- */
-public interface VMEventListener {
-
-    /**
-     * Receive an event from the VM. Events are sent in batches. See
-     * documentation of EventSet for more information.
-     *
-     * @param es Set of events
-     */
-    void vmEvent(EventSet es);
-}
